+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New use case `Redeploying dataflow on webhook` in docs.
 - Documentation for creating Source for new File types taking `.ini` as an example.
 - New input modes, output modes for HTTP API dataflow registration.
+- Usage example for tfhub text classifier.
 ### Changed
 - `Edit on Github` button now hidden for plugins.
 - Doctests now run via unittests